LensFactory.Cosmology — Module

Cosmology

Testing :-)

Module Index

LensFactory.Cosmology.init_cosmology

LensFactory.Cosmology.Ez

LensFactory.Cosmology.Omega_kz

LensFactory.Cosmology.Omega_mz

LensFactory.Cosmology.Omega_rz

LensFactory.Cosmology.Omega_wz

LensFactory.Cosmology.age

LensFactory.Cosmology.angular_scale

LensFactory.Cosmology.comoving_volume

LensFactory.Cosmology.comoving_volume_element

LensFactory.Cosmology.distance_modulus

LensFactory.Cosmology.hubble_distance

LensFactory.Cosmology.hubble_parameter

LensFactory.Cosmology.hubble_time

LensFactory.Cosmology.lookback_time

LensFactory.Cosmology.rho_cz

LensFactory.Cosmology.scale_factor

Initialization

LensFactory.Cosmology.init_cosmology — Type

init_cosmology(H0::RV=70.0, w::RV=-1.0, Omega_m0::RV=0.3, Omega_r0::RV=0.0, Omega_w0::RV=0.7)

An Abstract type to initialize a given cosmology.

Basics

LensFactory.Cosmology.scale_factor — Function

scale_factor(z::RV)::RV

Calculates the scale factor, $a$, at redshift, $z$, which is given as

\[a = 1 / (1+z)\]

LensFactory.Cosmology.Ez — Function

Ez(z::RV)::RV

Calculates the dimensionless Hubble parameter, $E(z)$, at redshift, $z$, which is given as,

\[E(z) = \sqrt{ \Omega_{m0} (1+z)^2 + \Omega_{r0} (1+z)^4 + \Omega_{k0} (1+z)^2 + \Omega_{w0} (1+z)^{3(1+w)} }\]

LensFactory.Cosmology.hubble_parameter — Function

hubble_parameter(cosmology::AbstractCosmology, z::RV)::RV

Calculates the Hubble parameter, $H(z) = H_0\:E(z)$, at redshift $z$

Time

LensFactory.Cosmology.hubble_time — Function

hubble_time(H0::RV)::RV

Calculates the Hubble time, $t_H$, (in Gyr), which is given as,

\[t_H = 1 / {\rm H_0}\]

LensFactory.Cosmology.age — Function

age(cosmo::AbstractCosmology, z::RV)::RV

Calculates the age, $t_{\rm age}$, (in Gyr) of the Universe at redshift, $z$, which is given as,

\[t_{\rm age}(z) = \int_z^\infty \frac{1}{(1+z')H(z')} dz' \]

LensFactory.Cosmology.lookback_time — Function

lookback_time(cosmology::AbstractCosmology, z::RV)::RV

Calculates lookbak time, $t_L(z)$ for a given redshift $z$, which is given as,

\[t_L(z) = \int_0^z \frac{1}{(1+z')H(z')} dz' \]

Density parameters

LensFactory.Cosmology.rho_cz — Function

rho_cz(cosmology, z)

Calculates the critical density of the Universe at redshift $z$

Formula: $\rho_c(z) = \frac{3 H^2(z)}{8 \pi {\rm G} }$`

Arguments

cosmology::AbstractCosmology: Cosmology instance

z::Union{Int64, Float64}: Input redshift

Returns

Union{Int64, Float64}: Critical density at redshift $z$

LensFactory.Cosmology.Omega_mz — Function

Omega_mz(cosmology, z)

Calculates the matter density parameter at redshift z.

Formula: $\Omega_{mz} = \Omega_{m0} (1+z)^3 \left( \frac{H0}{H(z)} \right)^2$

Arguments

cosmology::AbstractCosmology: Cosmology instance

z::Union{Int64, Float64}: Input redshift

Returns

Union{Int64, Float64}: Matter density parameter at redshift $z$

LensFactory.Cosmology.Omega_rz — Function

Omega_rz(cosmology, z)

Calculates the radiation density parameter at redshift $z$

Formula: $\Omega_{rz} = \Omega_{r0} (1+z)^4 \left( \frac{H_0}{H(z)} \right)^2$

Arguments

cosmology::AbstractCosmology: Cosmology instance

z::Union{Int64, Float64}: Input redshift

Returns

Union{Int64, Float64}: Radiation density parameter at redshift z

LensFactory.Cosmology.Omega_wz — Function

Omega_wz(cosmology, z)

Calculates the dark energy density parameter at redshift $z$

Formula: $\Omega_{wz} = \Omega_{w0} (1+z)^{3(1+w)} \left( \frac{H_0}{H(z)} \right)^2$

Arguments

cosmology::AbstractCosmology: Cosmology instance

z::Union{Int64, Float64}: Input redshift

Returns

Union{Int64, Float64}: Dark energy density parameter at redshift $z$

LensFactory.Cosmology.Omega_kz — Function

Omega_kz(cosmology, z)

Calculates the curvature density parameter at redshift z.

Formula: $\Omega_{kz} = \Omega_{k0} (1+z)^2 \left( \frac{H_0}{H(z)} \right)^2$

Arguments

cosmology::AbstractCosmology: Cosmology instance

z::Union{Int64, Float64}: Input redshift

Returns

Union{Int64, Float64}: Curvature density parameter at redshift $z$

Distances

LensFactory.Cosmology.hubble_distance — Function

hubble_distance(H0)

Calculates radius of the observable Universe.

Formula: $R_H = c/{\rm H_0}$

Arguments

H0::Union{Int64, Float64}: Hubble constant, in km/s/Mpc

Returns

Union{Int64, Float64}: Radius of the observable Universe, in meters

LensFactory.Cosmology.distance_modulus — Function

distance_modulus(cosmology, z)

Calculates the distance modulus to a given redshift $z$

Formula: $\mu = 5 \log(d_L/{\rm pc}) - 5$

Arguments

cosmology::AbstractCosmology: Cosmology instance

z::Union{Int64, Float64}: Input redshift

Returns

Union{Int64, Float64}: Distance modulus to a redshift $z$

LensFactory.Cosmology.angular_scale — Function

angular_scale(cosmo::AbstractCosmology, zf::RV)::RV

Volumes

LensFactory.Cosmology.comoving_volume_element — Function

comovingvolumeelement(cosmology, z)

Calculates the comving volume element at redshift z, in Gpc^3.

Formula: $ dV = dH \frac{D_M^2}{E(z)} $

Arguments

cosmology::AbstractCosmology: Cosmology instance

z::Union{Int64, Float64}: Input redshift

Returns

Union{Int64, Float64}: Comoving volume element at redshift z, in Gpc^3

LensFactory.Cosmology.comoving_volume — Function

comoving_volume(cosmology, z)

Calculates the comving volume element at redshift z

Formula: $ dV = dH \frac{D_M^2}{E(z)} $

Arguments

cosmology::AbstractCosmology: Cosmology instance

z::Union{Int64, Float64}: Redshift

Returns

Union{Int64, Float64}: Comoving volume element at redshift z, in Gpc^3