LensFactory.Lenses.init_ExternalEffects — Type

init_ExternalEffects(kappa::RV=NaN, gamma::RV=NaN, angle::RV=NaN)

Initialize constant external effects with the given parameters.

Keyword Arguments

kappa::RV = NaN: Convergence (dimensionless).

gamma::RV = NaN: Shear amplitude (dimensionless).

angle::RV = NaN: Shear angle (in $\rm \mathbf{degrees}$).

LensFactory.Lenses.ExternalEffects.potential! — Function

potential!(ψ::T, θx::T, θy::T, κ::RV, γ::RV, ϕ::RV) where T <: RV

potential!(ψ::T, θx::T, θy::T, κ::RV, γ::RV, ϕ::RV) where T <: ROA

Calculate potential at given coordinates for constant external convergence and shear and update the potential in place.

Arguments

ψ: Potential at given coordinates

θx: x-coordinate(s) (in $\rm \mathbf{arcseconds}$).

θy: y-coordinate(s) (in $\rm \mathbf{arcseconds}$).

κ::RV: External convergence.

γ::RV: External shear value.

ϕ::RV: External Shear angle (in $\rm \mathbf{degrees}$).

LensFactory.Lenses.ExternalEffects.deflection! — Function

deflection!(ψx::T, ψy::T, θx::T, θy::T, κ::RV, γ::RV, ϕ::RV) where T <: RV

deflection!(ψx::T, ψy::T, θx::T, θy::T, κ::RV, γ::RV, ϕ::RV) where T <: ROA

Calculate deflection at given coordinates for constant external convergence and shear and update the deflection in place.

Arguments

ψx: x-component of deflection at given coordinates

ψy: y-component of deflection at given coordinates

θx: x-coordinate(s) (in $\rm \mathbf{arcseconds}$).

θy: y-coordinate(s) (in $\rm \mathbf{arcseconds}$).

κ::RV: External convergence.

γ::RV: External shear value.

ϕ::RV: External Shear angle (in $\rm \mathbf{degrees}$).

LensFactory.Lenses.ExternalEffects.jacobian! — Function

jacobian!(ψxx::T, ψyy::T, ψxy::T, θx::T, θy::T, κ::RV, γ::RV, ϕ::RV) where T <: RV

jacobian!(ψxx::T, ψyy::T, ψxy::T, θx::T, θy::T, κ::RV, γ::RV, ϕ::RV) where T <: ROA

Calculate Jacobian at given coordinates for constant external convergence and shear and update the Jacobian in place.

Arguments

ψxx: xx-component of Jacobian at given coordinates

ψyy: yy-component of Jacobian at given coordinates

ψxy: xy-component of Jacobian at given coordinates

θx: x-coordinate(s) (in $\rm \mathbf{arcseconds}$).

θy: y-coordinate(s) (in $\rm \mathbf{arcseconds}$).

κ::RV: External convergence.

γ::RV: External shear value.

ϕ::RV: External Shear angle (in $\rm \mathbf{degrees}$).