Lenses

This module

LensFactory.Lenses.get_meshgrid — Function

get_meshgrid(θx::RV, θy::RV, dθ::RV) --> Tuple{Matrix{Float64}, Matrix{Float64}}

Generate a meshgrid of coordinates on which various quantities can be evaluated. At present, this function only generates square pixels. In future if the need arises, it can be extended to generate rectangular pixels as well.

Input:

θx::RV: Half-size of the grid in x-direction

θy::RV: Half-size of the grid in y-direction

dθ::RV: Pixel size

Output:

grid_x::Matrix{Float64}: x-coordinates of the grid

grid_y::Matrix{Float64}: y-coordinates of the grid

LensFactory.Lenses.get_critical_density — Function

get_critical_density(D_d::Float64, D_ds::Float64, D_s::Float64s; unit::Symbol=:kg_m2) --> Float64s

Calculate the critical surface density,

\[Σ_{\rm cr} = \frac{c^2}{4 π {\rm G}} \frac{D_s}{D_d D_{ds}},\]

given the angular diameter distances.

Input:

D_d::Float64: Angular diameter distance to the lens (in Mpc)

D_ds::Float64: Angular diameter distance to the source (in Mpc)

D_s::Float64: Angular diameter distance to the lens (in Mpc)

unit::Symbol: Unit of the result (default is :kg_m2)

:kg_m2 $\Rightarrow{\rm kg/m^2}$,

:msun_pc2 $\Rightarrow{\rm M_⊙/pc^2}$,

:msun_arcsec2 $\Rightarrow{\rm M_⊙/arcsec^2}$

Output:

Σ_cr::Float64: Critical surface density in the requested unit

LensFactory.Lenses.get_potential — Function

get_potential(lens::AbstractLens, θx::T, θy::T) where T <: RV --> RV

get_potential(lens::AbstractLens, θx::T, θy::T) where T <: ROA --> ROA

Calculate lensing potential at the given angular coordinates for the given lens model,

\[ψ(\pmb{θ}) = \frac{4{\rm G}}{\rm c^2} \frac{1}{D_d} \int d^2 \pmb{θ}' \, Σ(\pmb{θ}') \, \ln\left(|\pmb{θ} - \pmb{θ'}|\right).\]

Input:

lens::AbstractLens: Lens model

θx::T: x-coordinate(s)

θy::T: y-coordinate(s)

Output:

ψ::ROA: Lensing potential at the given angular coordinate(s)

LensFactory.Lenses.get_deflection — Function

get_deflection(lens::AbstractLens, θx::T, θy::T) where T <: RV --> Tuple{RV, RV}

get_deflection(lens::AbstractLens, θx::T, θy::T) where T <: ROA --> Tuple{ROA, ROA}

Calculate (vector) deflection angle at the given angular coordinate(s) for a given lens model,

\[\pmb{α}(\pmb{θ}) = \pmb{∇} ψ(\pmb{θ}) = \frac{4{\rm G}}{\rm c^2} \frac{1}{D_d} \int d^2 \pmb{θ}' \, Σ(\pmb{θ}') \frac{\pmb{θ} - \pmb{θ}'}{|\pmb{θ} - \pmb{θ}'|^2}.\]

Input:

lens::AbstractLens: Lens model

θx::T: x-coordinate(s)

θy::T: y-coordinate(s)

Output:

αx::ROA: x-component of the deflection angle at the given angular coordinate(s)

αy::ROA: y-component of the deflection angle at the given angular coordinate(s)

LensFactory.Lenses.get_jacobian — Function

get_jacobian(lens::AbstractLens, θx::T, θy::T) where T <: RV --> Tuple{RV, RV, RV}

get_jacobian(lens::AbstractLens, θx::T, θy::T) where T <: ROA --> Tuple{ROA, ROA, ROA}

Calculate jacobian (i.e., deformation tensor) of the lens mapping for a given lens model,

\[\mathcal{A}(\pmb{θ}) = \begin{pmatrix} ψ_{xx} & ψ_{xy} \\ ψ_{xy} & ψ_{yy} \end{pmatrix}.\]

Since the jacobian is symmetric (for single lens plane), only three components are returned, i.e., $(ψ_{xx}, ψ_{yy}, ψ_{xy})$.

Input:

lens::AbstractLens: Lens model

θx::T: x-coordinate(s)

θy::T: y-coordinate(s)

Output:

ψxx::ROA: xx-component of the jacobian at the given angular coordinate(s)

ψyy::ROA: yy-component of the jacobian at the given angular coordinate(s)

ψxy::ROA: xy-component of the jacobian at the given angular coordinate(s)

LensFactory.Lenses.get_time_delay — Function

get_time_delay(lens::AbstractLens, θx::T, θy::T, adis::Float64, z_d::RV, D_d::RV, β::NTuple{2, RV}) where T <: RV --> RV

get_time_delay(lens::AbstractLens, θx::T, θy::T, adis::Float64, z_d::RV, D_d::RV, β::NTuple{2, RV}) where T <: ROA --> ROA

Calculate time delay for a given lens model and source position. The corresponding expression is given as,

\[t_d(\pmb{θ}; \pmb{β}) = \frac{1+z_l}{\rm c} \frac{D_d D_s}{D_{ds}} \theta_0^2 \left[ \frac{(\pmb{θ} - \pmb{β})^2}{2} - \frac{D_{ds}}{D_s} \psi(\pmb{θ}) \right],\]

where $\theta_0$ is normalizing angular unit.

Input:

lens::AbstractLens: Lens model

θx::T: x-coordinate(s)

θy::T: y-coordinate(s)

adis::Float64: Distance ratio (i.e., $D_{ds}/D_s$)

z_d::RV: Lens redshift

D_d::RV: Angular diameter distance to the lens (in meters)

β::NTuple{2, RV}: Source angular position (in arcseconds)

Output:

t_d::ROA: Time delay at the given angular coordinate(s)

LensFactory.Lenses.get_magnification_image — Function

get_magnification_image(lens::AbstractLens, θx::T, θy::T) where T <: RV --> RV

get_magnification_image(lens::AbstractLens, θx::T, θy::T) where T <: ROA --> ROA

Calculate signed magnification at the given angular coordinate(s) for a given lens model,

\[\mu(\pmb{θ}) = \frac{1}{det\left[ \mathbb{I} - \mathcal{A} \right]} = \frac{1}{1 + ψ_{xx} ψ_{yy} - ψ_{xx} - ψ_{yy} - ψ_{xy}^2},\]

where $\mathbb{I}$ is the identity matrix.

Input:

lens::AbstractLens: Lens model

θx::T: x-coordinate(s)

θy::T: y-coordinate(s)

adis::Float64: Distance ratio (i.e., $D_{ds}/D_s$)

Output:

μ::ROA: Magnification at the given angular coordinate(s)

LensFactory.Lenses.get_magnification_source — Function

get_magnification_source(lens::AbstractLens, θx::T, θy::T, adis::Float64; rays_per_pixel::Int64=1) where T <: Matrix{<:RV} --> Matrix{RV}

Calculates the magnification map in source plane using inverse ray shooting (IRS) for a given lens model. The number of average rays per pixel can be specified using the rays_per_pixel keyword argument. This function is not optimized for speed and is only intended to visualize the magnification map.

Input:

lens::AbstractLens: Lens model

θx::Matrix{<:RV}: x-grid

θy::Matrix{<:RV}: y-grid

adis::Float64: Distance ratio (i.e., $D_{ds}/D_s$)

rays_per_pixel::Int64: Average number of rays per pixel

Output:

μ_source::Matrix{<:RV}: Magnification map in source plane

LensFactory.Lenses.get_image — Function

get_image(lens::AbstractLens, θx::ROA, θy::ROA, adis::Float64, β::NTuple{2, RV}) --> Vector{NTuple{2, RV}}

Calculate image positions for a given lens model and source position. To get the image positions, this implementation finds the intersection points of contours corresponding to,

\[\pmb{β} - \pmb{θ} + a_{\rm dis} \, \pmb{α}(\pmb{θ}) = 0,\]

where $\pmb{β}$ is the source position, $\pmb{θ}$ is the image plane grid, $a_{\rm dis}$ is the distance ratio (i.e., $D_{ds}/D_s$), and $\pmb{α}(\pmb{θ})$ is the deflection angle. To find the intersection points inside the pixels, we use bi-linear interpolation.

Input:

lens::AbstractLens: Lens model

θx::ROA: x-grid

θy::ROA: y-grid

adis::Float64: Distance ratio (i.e., $D_{ds}/D_s$)

β::NTuple{2, RV}: Source position

Output:

image_position::Vector{NTuple{2, RV}}: Image positions

LensFactory.Lenses.get_critical_curve — Function

get_critical_curve(lens::AbstractLens, θx::T, θy::T, adis::Float64) where T <: Matrix{<:RV} --> Tuple{Vector{Vector{Vector{Float64}}}, Vector{Vector{Vector{Float64}}}}

Calculates the critical curves for a given lens model.

LensFactory.Lenses.get_caustic — Function

get_caustic(lens::AbstractLens, θx::T, θy::T, adis::Float64) where T <: Matrix{<:RV}

LensFactory.Lenses.get_critical_area — Function

get_critical_area(lens::AbstractLens, θx::T, θy::T, adis::Float64) where T <: Matrix{<:RV} --> Float64

LensFactory.Lenses.get_einstein_angle — Function

get_einstein_angle(lens::AbstractLens, θx::T, θy::T, adis::Float64) where T <: Matrix{<:RV} --> Float64

LensFactory.Lenses.shear_cartesian2polar — Function

shear_cartesian2polar(γ1::T, γ2::T) where T <: RV --> Tuple{Float64, Float64}

LensFactory.Lenses.shear_polar2cartesian — Function

shear_polar2cartesian(γ::T, phi::T) where T <: RV --> Tuple{Float64, Float64}

LensFactory.Lenses.ellipticity_cartesian2polar — Function

ellipticity_cartesian2polar(e1::T, e2::T) where T <: RV --> Tuple{Float64, Float64}

LensFactory.Lenses.ellipticity_polar2cartesian — Function

ellipticity_polar2cartesian(e::T, phi::T) where T <: RV --> Tuple{Float64, Float64}