This part contains plotting support for the LensFactory.jl package.

LensFactory.Lenses.plot_image_plane — Function

LensFactory.Lenses.plot_image_plane(lens::Lenses.AbstractLens, θx::ROA, θy::ROA, adis::RV)

Additional keyword arguments and their default values

two_panel::Bool = false – Whether to create a two-panel plot with source plane on the left and image plane on the right

plot_caustic::Bool = true

caustic_kws::NamedTuple = (color_tan = :green, color_rad = :green, linewidth = 2)

plot_critical::Bool =true

critical_kws::NamedTuple = (color_tan = :red, color_rad = :red, linewidth = 2)

source::Union{Nothing, NTuple{2, RV}, Matrix{<:RV}} = nothing

source_kws::NamedTuple = (color=:red, markersize=10, marker=:star5, heatmap=cgrad([:white, :blue]))

image_kws::NamedTuple = (color=:blue, markersize=10, marker=:star5, heatmap=cgrad([:white, :red]))

save_plot::Bool = false

plot_name::String = "image_plane.png"

resolution::Int = 2

LensFactory.Lenses.plot_surface_density — Function

LensFactory.Lenses.plot_surface_density(lens::Lenses.AbstractLens, θx::ROA, θy::ROA, adis::Float64; D_d::Float6=NaN)

Additional keyword arguments and their default values

unit::Symbol = :kg_m2 – Unit for surface density. Options are :convergence, :kgm2, :msunpc2, :msun_arcsec2

figure_size::NTuple{2, RV} = (500, 400)

heatmap_kws::NamedTuple = (colormap=:cubehelix, colorrange=(0, 6))

plot_contour::Bool = false

contour_kws::NamedTuple = (levels=0.5:0.2:1.5, labels=false)

save_plot::Bool = false

plot_name::String = "surface_density.png"

resolution::Int = 2

LensFactory.Lenses.plot_magnification_map — Function

LensFactory.Lenses.plot_magnification_map(lens::Lenses.AbstractLens, θx::ROA, θy::ROA, adis::Float64)

Additional keyword arguments and their default values

plane::Symbol = :image – Whether to plot |μ| map in :image or :source plane

rays_per_pixel::Int64 = 1 – Only relevant if plane = :source

figure_size::NTuple{2, RV} = (500, 400)

heatmap_kws::NamedTuple = (colormap=:binary, colorrange=(1, 100))

save_plot::Bool = false

plot_name::String = "magnification_map.png"

resolution::Int = 2

LensFactory.Lenses.plot_magnification_profile — Function

LensFactory.Lenses.plot_magnification_profile(lens::Lenses.AbstractLens, θx::ROA, θy::ROA, adis::Float64)

Additional keyword arguments and their default values

plane::Symbol = :image – Whether to plot |μ| map in :image or :source plane

rays_per_pixel::Int64 = 1 – Only relevant if plane = :source

mu_range::StepRange{<:RV, <:RV} = 1:5:500

unit::Float64 = ANGLE_ARCMIN^2

figure_size::NTuple{2, RV} = (500, 400)

plot_kws::NamedTuple = (color=:black, linewidth=2, linestyle=:solid)

save_plot::Bool = false

plot_name::String = "magnification_profile.png"

resolution::Int = 2

LensFactory.MultiPlane.plot_image_plane — Function

LensFactory.MultiPlane.plot_image_plane(cosmology::Cosmology.AbstractCosmology, lens::Lenses.AbstractLens, θx::ROA, θy::ROA, zs::RV)

Additional keyword arguments and their default values

two_panel::Bool = false – Whether to create a two-panel plot with source plane on the left and image plane on the right

plot_caustic::Bool = true

caustic_kws::NamedTuple = (color = :green, linewidth = 2)

plot_critical::Bool =true

critical_kws::NamedTuple = (color = :red, linewidth = 2)

source::Union{Nothing, NTuple{2, RV}, Matrix{<:RV}} = nothing

source_kws::NamedTuple = (color=:red, markersize=10, marker=:star5, heatmap=cgrad([:white, :blue]))

image_kws::NamedTuple = (color=:blue, markersize=10, marker=:star5, heatmap=cgrad([:white, :red]))

save_plot::Bool = false

plot_name::String = "image_plane.png"

resolution::Int = 2