LensFactory.LensModel.plot_corner — Function

LensFactory.LensModel.plot_corner(chains, lls; param_names=nothing, burn_in=0.3, thinning=100)

Generates a corner plot for the given MCMC chains and log-likelihoods.

Arguments

chains: MCMC chains of shape (nsteps, nchains, n_params).

lls: Log-likelihoods corresponding to the chains, of shape (nsteps, nchains).

Keyword arguments

param_names=nothing: Optional list of parameter names for labeling the axes.

burn_in=0.3: Fraction of the initial samples to discard as burn-in.

thinning=100: Interval for thinning the chains to reduce autocorrelation.

Returns

A Makie figure object containing the corner plot.

LensFactory.LensModel.plot_trace — Function

LensFactory.LensModel.plot_trace(chains; param_names=nothing, burn_in=0.0, thinning=1)

Generates trace plots for the given MCMC chains.

Arguments

chains: MCMC chains of shape (nsteps, nchains, n_params).

Keyword arguments

param_names=nothing: Optional list of parameter names for labeling the axes.

burn_in=0.0: Fraction of the initial samples to discard as burn-in.

thinning=1: Interval for thinning the chains to reduce autocorrelation.

Returns

A Makie figure object containing the trace plots.

LensFactory.LensModel.plot_best_model — Function

LensFactory.LensModel.plot_best_model(θx, θy, model, chains, lls)

Generates a plot comparing the observed image positions with the predicted image positions from the best-fit lens model.

Arguments

model: The lens model configuration.

chains: MCMC chains of shape (nsteps, nchains, n_params).

lls: Log-likelihoods corresponding to the chains, of shape (nsteps, nchains).

Returns

A Makie figure object containing the comparison plot.