Point Lens

Arguably, the simplest gravitational lens is a point mass lens (i.e., Schwarzschild lens). The lensing by a point mass lens is characterized by two parameters: its mass ($M$) and the source position ($\boldsymbol{\beta}$).

In LensFactory, to define a point lens, the user need to specify four parameters: its position in the image plane ($x_c, y_c$), its mass ($M$) and the angular diameter distance to the lens ($D_d$). By default, the lens is placed at the origin, i.e., ($x_c, y_c$) = (0,0). From above, to initialize a point lens, we need to first define a cosmology and calculate the angular diameter distance to the lens. We refer reader to Example - 2 for more details.

LensFactory.Lenses.init_PointLens — Type

init_PointLens(D_d::RV=NaN, x_c::RV=0.0, y_c::RV=0.0, mass::RV=NaN)

Initialize a point lens with the given parameters.

Keyword Arguments

D_d::RV = NaN: ADD from observer to lens (in $\rm \mathbf{meters}$).

x_c::RV = 0.0: x-coordinate of the lens (in $\rm \mathbf{arcseconds}$).

y_c::RV = 0.0: y-coordinate of the lens (in $\rm \mathbf{arcseconds}$).

mass::RV= NaN: Mass of the lens (in $\rm \mathbf{M_\odot}$).

LensFactory.Lenses.PointLens.potential! — Function

potential!(ψ::T, θx::T, θy::T, D_d::RV, θxc::RV, θyc::RV, mass::RV) where T <: RV

potential!(ψ::T, θx::T, θy::T, D_d::RV, θxc::RV, θyc::RV, mass::RV) where T <: ROA

Calculate potential at given coordinates for a point mass lens and update the potential (ψ) in place. The lensing potential is given as,

\[\psi(θ_x, θ_y) = \frac{4 \, \rm{G} \, M}{\rm{c}^2} \frac{1}{D_d} \ln |\pmb{θ} - \pmb{θ}_c|.\]

Arguments

ψ: Potential at given coordinates

θx: x-coordinate(s) (in $\rm \mathbf{arcseconds}$).

θy: y-coordinate(s) (in $\rm \mathbf{arcseconds}$).

D_d::RV: ADD from the observer to the lens (in $\rm \mathbf{meters}$).

θxc::RV: x-coordinate of the lens (in $\rm \mathbf{arcseconds}$).

θyc::RV: y-coordinate of the lens (in $\rm \mathbf{arcseconds}$).

mass::RV: Mass of the lens (in $\rm \mathbf{M_\odot}$).

Returns

nothing: Updates the potential (ψ) in place.

LensFactory.Lenses.PointLens.deflection! — Function

deflection!(ψx::T, ψy::T, θx::T, θy::T, D_d::RV, θxc::RV, θyc::RV, mass::RV) where T <: RV

deflection!(ψx::T, ψy::T, θx::T, θy::T, D_d::RV, θxc::RV, θyc::RV, mass::RV) where T <: ROA

Calculate deflection at given coordinates for a point mass lens and update the deflection components (ψx, ψy) in place. The deflection angle components are given as,

\[\begin{align*} ψ_x(θ_x, θ_y) &= \frac{4 \, \rm{G} \, M}{\rm{c}^2} \frac{1}{D_d} \frac{θ_x - θ_{x,c}}{|\pmb{θ} - \pmb{θ}_c|^2}, \\ ψ_y(θ_x, θ_y) &= \frac{4 \, \rm{G} \, M}{\rm{c}^2} \frac{1}{D_d} \frac{θ_y - θ_{y,c}}{|\pmb{θ} - \pmb{θ}_c|^2}. \end{align*}\]

Arguments

ψx: x-component of the deflection at given coordinates

ψy: y-component of the deflection at given coordinates

θx: x-coordinate(s) (in $\rm \mathbf{arcseconds}$).

θy: y-coordinate(s) (in $\rm \mathbf{arcseconds}$).

D_d::RV: ADD from the observer to the lens (in $\rm \mathbf{meters}$).

θxc::RV: x-coordinate of the lens (in $\rm \mathbf{arcseconds}$).

θyc::RV: y-coordinate of the lens (in $\rm \mathbf{arcseconds}$).

mass::RV: Mass of the lens (in $\rm \mathbf{M_\odot}$).

Returns

nothing: Updates the deflection (ψx, ψy) in place.

LensFactory.Lenses.PointLens.jacobian! — Function

jacobian!(ψxx::T, ψyy::T, ψxy::T, θx::T, θy::T, D_d::RV, θxc::RV, θyc::RV, mass::RV) where T <: RV

jacobian!(ψxx::T, ψyy::T, ψxy::T, θx::T, θy::T, D_d::RV, θxc::RV, θyc::RV, mass::RV) where T <: ROA

Calculate jacobian at given coordinates for a point mass lens and update the jacobian components (ψxx, ψyy, ψxy) in place. The jacobian components are given as,

\[\begin{align*} \psi_{xx} &= \frac{4 \, \rm{G} \, M}{\rm{c}^2} \frac{1}{D_d} \frac{(θ_y - θ_{y,c})^2 - (θ_x - θ_{x,c})^2}{|\pmb{θ} - \pmb{θ}_c|^4}, \\ \psi_{yy} &= \frac{4 \, \rm{G} \, M}{\rm{c}^2} \frac{1}{D_d} \frac{(θ_x - θ_{x,c})^2 - (θ_y - θ_{y,c})^2}{|\pmb{θ} - \pmb{θ}_c|^4}, \\ \psi_{xy} &= \frac{4 \, \rm{G} \, M}{\rm{c}^2} \frac{1}{D_d} \frac{-2 \, (θ_x - θ_{x,c}) (θ_y - θ_{y,c})}{|\pmb{θ} - \pmb{θ}_c|^4}. \end{align*}\]

Arguments

ψxx: x-component of the jacobian at given coordinates

ψyy: y-component of the jacobian at given coordinates

ψxy: xy-component of the jacobian at given coordinates

θx: x-coordinate(s) (in $\rm \mathbf{arcseconds}$).

θy: y-coordinate(s) (in $\rm \mathbf{arcseconds}$).

D_d::RV: ADD from the observer to the lens (in $\rm \mathbf{meters}$).

θxc::RV: x-coordinate of the lens (in $\rm \mathbf{arcseconds}$).

θyc::RV: y-coordinate of the lens (in $\rm \mathbf{arcseconds}$).

mass::RV: Mass of the lens (in $\rm \mathbf{M_\odot}$).

Returns

nothing: Updates the jacobian (ψxx, ψyy, ψxy) in place.

LensFactory.Lenses.PointLens.einstein_angle — Function

einstein_angle(;D_d::Float64=NaN, D_ds::Float64=NaN, D_s::Float64=NaN, mass::Float64=NaN)

Calculate the Einstein angle for a point mass lens,

\[\theta_E = \sqrt{\frac{4 \, \rm{G} \, M}{\rm{c}^2} \frac{D_{ds}}{D_d D_s}}.\]

Keyword Arguments

D_d::Float64=NaN: ADD from observer to lens (in $\rm \mathbf{meters}$).

D_ds::Float64=NaN: ADD from lens to source (in $\rm \mathbf{meters}$).

D_s::Float64=NaN: ADD from observer to source (in $\rm \mathbf{meters}$).

mass::Float64=NaN: Mass of the lens (in $\rm \mathbf{M_\odot}$).

Returns

θE: Einstein angle (in $\rm \mathbf{arcseconds}$)