LensFactory.Lenses.init_aNFWLens — Type

init_aNFWLens(cosmology::AbstractCosmology, z_d::Real; 
              x_c::Real  = 0.0, 
              y_c::Real  = 0.0, 
              mass::Real = NaN, 
              x_s::Real  = NaN, 
              c::Real    = NaN, 
              eps::Real  = NaN, 
              pa::Real   = NaN)

Initialize an approximate Navarro-Frenk-White lens (aNFWLens) with the given parameters based on Oguri (2021). The lens model can be initialized with either the concentration c or the scale radius x_s. If both are provided, c will be used to calculate x_s and the input x_s will be overwritten.

Arguments

cosmology::AbstractCosmology: Cosmology object.

z_d::Real: Redshift of the lens.

Keyword Arguments

x_c::Real = 0.0: x-coordinate of the lens (in $\rm \mathbf{arcseconds}$).

y_c::Real = 0.0: y-coordinate of the lens (in $\rm \mathbf{arcseconds}$).

mass::Real= NaN: Mass of the lens (in $\rm \mathbf{M_\odot}$).

x_s::Real = NaN: Scale radius (in $\rm \mathbf{arcseconds}$).

c::Real = NaN: Concentration of the lens.

eps::Real = NaN: Ellipticity.

pa::Real = NaN: Position angle (in $\rm \mathbf{deg}$).

LensFactory.Lenses.aNFWLens.potential! — Function

potential!(ψ::U, θx::S, θy::S, D_d::T, θxc::T, θyc::T, k_s::T, θs::T, ϵ::T, pa::T) where {U<:Real, S<:Real, T<:Real}

potential!(ψ::U, θx::S, θy::S, D_d::T, θxc::T, θyc::T, k_s::T, θs::T, ϵ::T, pa::T) where {U<:ROA, S<:ROA, T<:Real}

LensFactory.Lenses.aNFWLens.deflection! — Function

deflection!(ψx::U, ψy::U, θx::S, θy::S, D_d::T, θxc::T, θyc::T, k_s::T, θs::T, ϵ::T, pa::T) where {U<:Real, S<:Real, T<:Real}

deflection!(ψx::U, ψy::U, θx::S, θy::S, D_d::T, θxc::T, θyc::T, k_s::T, θs::T, ϵ::T, pa::T) where {U<:ROA, S<:ROA, T<:Real}

LensFactory.Lenses.aNFWLens.jacobian! — Function

jacobian!(ψxx::U, ψyy::U, ψxy::U, θx::S, θy::S, D_d::T, θxc::T, θyc::T, k_s::T, θs::T, ϵ::T, pa::T) where {U<:Real, S<:Real, T<:Real}

jacobian!(ψxx::U, ψyy::U, ψxy::U, θx::S, θy::S, D_d::T, θxc::T, θyc::T, k_s::T, θs::T, ϵ::T, pa::T) where {U<:ROA, S<:ROA, T<:Real}