LensFactory.Cosmology — Module

Cosmology

Testing :-)

Methods

LensFactory.Cosmology.init_cosmology — Type

init_cosmology

An Abstract type to initialize cosmology.

Arguments

H0::Union{Int64, Float64} = 70: Hubble constant, in km/s/Mpc

w::Union{Int64, Float64} = -1.0: Dark energy EOS parameter

Ω_m0::Union{Int64, Float64} = 0.3: Present matter density parameter

Ω_r0::Union{Int64, Float64} = 0.0: Present radiation density parameter

Ω_w0::Union{Int64, Float64} = 0.7: Present dark energy density parameter

Returns

cosmology::AbstractCosmology: Cosmology instance defined by the input parameters

LensFactory.Cosmology.scale_factor — Function

scale_factor(z)

Calculates the scale factor at a given redshift.

Formula: $ a = \frac{1}{1+z} $

Arguments

z::Union{Int64, Float64}: Redshift

Returns

Union{Int64, Float64}: Scale factor

LensFactory.Cosmology.hubble_time — Function

hubble_time(H0)

Calculates age of the Universe (t_0).

Formula: $ t_0 = 1.0 / H0 $

Arguments

H0::Union{Int64, Float64}: Hubble constant, in km/s/Mpc

Returns

Union{Int64, Float64}: Age of the Universe, in Gyr

LensFactory.Cosmology.hubble_parameter — Function

hubble_parameter(cosmology, z)

Calculates Hubble parameter at redshift z.

Formula:

Arguments

cosmology::AbstractCosmology: Cosmology instance

z::Union{Int64, Float64}: Input redshift

Returns

Union{Int64, Float64}: Hubble parameter at redshift z, in km/s/Mpc

LensFactory.Cosmology.hubble_distance — Function

hubble_distance(H0)

Calculates radius of the observable Universe (R_H).

Formula: $\R_H = \frac{c}{H0}$

Arguments

H0::Union{Int64, Float64}: Hubble constant, in km/s/Mpc

Returns

Union{Int64, Float64}: Radius of the observable Universe, in meters

LensFactory.Cosmology.distance_modulus — Function

distance_modulus(cosmology, z)

Calculates the distance modulus to a given redshift z.

Formula: \mu = 5 \log(d) - 5

Arguments

cosmology::AbstractCosmology: Cosmology instance

z::Union{Int64, Float64}: Input redshift

Returns

Union{Int64, Float64}: Distance modulus to a redshift z

LensFactory.Cosmology.rho_cz — Function

rho_cz(cosmology, z)

Calculates the critical density of the Universe at redshift z.

Formula: \rho_c(z) = \frac{3 H^2(z)}{8\pi G}

Arguments

cosmology::AbstractCosmology: Cosmology instance

z::Union{Int64, Float64}: Input redshift

Returns

Union{Int64, Float64}: Critical density at redshift z

LensFactory.Cosmology.Omega_mz — Function

Omega_mz(cosmology, z)

Calculates the matter density parameter at redshift z.

Formula: Ω_mz = Ω_m0 * (1+z)^3 * (H0/H(z))^2

Arguments

cosmology::AbstractCosmology: Cosmology instance

z::Union{Int64, Float64}: Input redshift

Returns

Union{Int64, Float64}: Matter density parameter at redshift z

LensFactory.Cosmology.Omega_rz — Function

Omega_rz(cosmology, z)

Calculates the radiation density parameter at redshift z.

Formula: Ω_rz = Ω_r0 * (1+z)^4 * (H0/H(z))^2

Arguments

cosmology::AbstractCosmology: Cosmology instance

z::Union{Int64, Float64}: Input redshift

Returns

Union{Int64, Float64}: Radiation density parameter at redshift z

LensFactory.Cosmology.Omega_wz — Function

Omega_wz(cosmology, z)

Calculates the dark energy density parameter at redshift z.

Formula: Ω_wz = Ω_w0 * (1+z)^(3*(1+w)) * (H0/H(z))^2

Arguments

cosmology::AbstractCosmology: Cosmology instance

z::Union{Int64, Float64}: Input redshift

Returns

Union{Int64, Float64}: Dark energy density parameter at redshift z

LensFactory.Cosmology.Omega_kz — Function

Omega_kz(cosmology, z)

Calculates the curvature density parameter at redshift z.

Formula: Ω_kz = Ω_k0 * (1+z)^2 * (H0/H(z))^2

Arguments

cosmology::AbstractCosmology: Cosmology instance

z::Union{Int64, Float64}: Input redshift

Returns

Union{Int64, Float64}: Curvature density parameter at redshift z

LensFactory.Cosmology.lookback_time — Function

lookback_time(cosmology, z)

Calculates lookbak time for a given redshift $z$

Formula: $ tL(z) = \int0^z \frac{1}{(1+z')H(z')} dz' $

Arguments

cosmology::AbstractCosmology: Cosmology instance

z::Union{Int64, Float64}: Input redshift

Returns

Union{Int64, Float64}: Lookback time for a source at redshift $z$

LensFactory.Cosmology.comoving_volume_element — Function

comovingvolumeelement(cosmology, z)

Calculates the comving volume element at redshift z, in Gpc^3.

Formula: $ dV = dH \frac{D_M^2}{E(z)} $

Arguments

cosmology::AbstractCosmology: Cosmology instance

z::Union{Int64, Float64}: Input redshift

Returns

Union{Int64, Float64}: Comoving volume element at redshift z, in Gpc^3