LensModel

The LensModel module is the interface between the user and the lens modeling engine of LensFactory.jl. All user inputs are specified through a YAML file. The input YAML is divided into multiple sections. The overall structure of the input YAML file is given below.

observation:
   modeler: LensFactory
   lens: WhoKnows
   z_d: ...
   reference: [RA, DEC]
   pixel_scale: ...
   FOV: [x, y]

cosmology:
   H0: 70.0
   Omega_m0: 0.30
   Omega_r0: 0.00
   Omega_w0: 0.70

source:
   total_sources: S
   from_file: false
   source1:
      z_s:
         value: zv
         range: [zl, zu]
      total_knots: K
      knot1:
         x: [x1, x2, x3, x4]
         y: [y1, y2, y3, y4]
         sigma_x: [σx1, σx2, σx3, σx4]
         sigma_y: [σy1, σy2, σy3, σy4]
         sigma_theta: [σθ1, σθ2, σθ3, σθ4]
      .
      .
      .
      knotK:
         x: [x1, x2]
         y: [y1, y2]
         sigma_x: [σx1, σx2]
         sigma_y: [σy1, σy2]
         sigma_theta: [σθ1, σθ2]
   .
   .
   .
   sourceS:
      z_s: zs
      total_knots: 1
      knot1:
         x: [x1, x2, x3]
         y: [y1, y2, y3]
         sigma_x: [σx1, σx2, σx3]
         sigma_y: [σy1, σy2, σy3]
         sigma_theta: [σθ1, σθ2, σθ3]

lens_model:
   multiplane: false
   total_lenses: N
   lens1:
      lens: SIELens
      param1:
         value: p1v
         range: [p1l, p1u]
      param2:
         value: p2v
         range: [p2l, p2u]
      .
      .
      .
      paramN:
         value: p3v
         range: [p3l, p3u]
   .
   .
   .
   lensN:
      param1: ...
      param2: ...
      .
      .
      .

sampling:
   scheme: SourcePlane
   verbose: true
   optimizer:
   mcmc:

Looking at the above structure, we note that input YAML file is divided into five sections, i.e., observation, cosmology, source, lens_model, and sampling. Below, we describe each section in more detail. Overall, the parameters can be divided into three categories: optional, mandetory, and free. To make a parameter free, the user needs to specify initial value and range keywords for the same.

observation

The observation section contains the following keywords.

modeler: (optional) Modeler name (default: LensFactory).

lens: (optional) Lens name (default: WhoKnows).

z_d: (mandatory) Lens redshift.

reference: (mandatory) Reference [RA, DEC] position of the lens in degrees. If [0.0, 0.0] is provided then it would be assumed that the we are working with coordinates in arcseconds.

pixel_scale: (mandatory) Pixel scale in arcseconds.

FOV: (mandatory) Field of view in arcseconds. The user can either specify a single value (e.g., 20) or a vector of two values [x, y] (e.g., [10, 10]). If a single value is provided then square grid of [-value, value] will be used otherwise a rectangular grid with size [-x, x] and [-y, y] will be used.

cosmology

The cosmology section specifies the underlying cosmology and contains the following keywords. Each one of these parameters can be left free.

H0: (free) Hubble parameter (default: 70.0).

Omega_m0: (free) Matter density parameter (default: 0.30).

Omega_r0: (free) Radiation density parameter (default: 0.00).

Omega_w0: (free) Dark energy density parameter (default: 0.70).

Note

This section can be considered as an optional section. If no comoslogy is provided then the code will assume the default cosmology. However, it is highly recommended to explicity provide the cosmolgy.

source

The source section contains the details about the observed strongly lensed sources. The corresponding details can be specified in two ways, either directly in the YAML file or through a seperate .txt file. Each source will have an associated redshift and a number of knots.

total_sources: (mandatory) Total number of sources.

from_file: (optional) Boolean value indicating whether the source model is read from a file (default: false). If true, the source parameters will be read from the file specified in the source_file keyword.

source_file: (optional) Path to the .txt file containing the source parameters. This keyword is only required when from_file is set to true. For each source, the only possible free parameter is its redshifts.

source1: Details of the first source.

z_s: (free) Source redshift.

total_knots: (mandatory) Total number of strongly lensed knots within the source.

knot1: Details of the first knot.

x: (mandatory) x-coordinates of the knots.

y: (mandatory) y-coordinates of the knots.

sigma_x: (mandatory) x-component of error ellipse.

sigma_y: (mandatory) x-component of error ellipse.

sigma_theta: (mandatory) Angle of error ellipse (measure with respect to x-axis in anti-clockwise direction).

. . .

knotK: Details of knotK of source1.

. . .

sourceS: Details of sourceS.

The format of the source .txt file is as follows:

| sourceID | knotID | x | y | z_s | σx | σy | σθ | parity | magnitude | time delay |

Warning

The support for parity, magnitude, and time delay in modelling is under testing. Hence, the user should avoid using these keywords during lens modelling.

lens_model

The lens_model section contains the following keywords.

multiplane: (optional) Boolean value indicating whether the lens model is multiplane (default: false).

total_lenses: (mandatory) Total number of lenses.

lens1: (mandatory) Model for the first lens component.

lens: (mandatory) Lens name.

param1: Parameter-1.

value: initial value

range: [lower, upper]

param2: Parameter-2.

. . .

paramN: Parameter N (mandatory).

. . .

lensN: Model for the last lens.

sampling

The sampling section contains the following keywords.

scheme: (mandatory) Sampling scheme (default: SourcePlane). Currently implemented methods are SourcePlane.

verbose: (optional) Boolean value indicating whether the sampling should be verbose (default: true).

optimizer: Optimizer to be used for the sampling. The default optimizer is NM.

mcmc: Markov chain Monte Carlo parameters. The following keywords are available.

iterations: Number of iterations (default: 1000).

burn_in: Number of burn-in iterations (default: 500).

thin: Thinning factor (default: 1).

temperature: Temperature parameter (default: 1.0).

step_size: Step size (default: 1.0).

LensFactory.LensModel.read_input — Function

read_input(input_filename::AbstractString)

Reads the input YAML file and constructs a ModelConfig struct containing all the necessary information for lens modeling and sampling. For details on the expected structure of the input YAML file, please refer to Example - 2.

Arguments

input_filename::AbstractString: Path to the input YAML file.

Returns

ModelConfig::ModelConfig: A struct containing the observation details, cosmology, lens configuration, source configuration, parameter definitions, and sampling configuration.

LensFactory.LensModel.fit_model — Function

fit_model(model::ModelConfig; save::Bool=true, file_name::Union{String, Nothing}=nothing)

Performs lens model fitting using the given model.

Arguments

model::ModelConfig: The lens model configuration containing the observation, lens, source, and sampler details.

save::Bool=true: Whether to save the MCMC results to a JLD2 file (default: true)

file_name::Union{String, Nothing}=nothing: The name of the JLD2 file to save results. If no name is provided then a name will be generated based on the lens name and current date (i.e., LensName_DDMMYYYY.jld2).

Returns

chains::Array{Float64, 3}: The MCMC chains containing sampled parameter values

log_posterior::Matrix{Float64}: The log-posterior values corresponding to each sample in the chains

LensFactory.LensModel.free_parameter_names — Function

free_parameter_names(model::ModelConfig)

Get a vector of tuples containing the owner and name of each free parameter in the model.

Arguments

model::ModelConfig: The lens model configuration containing the observation, lens, source, and sampler details.

Returns

Vector{Tuple{Symbol, Symbol}}: A vector of tuples containing the owner and name of each free parameter.

LensFactory.LensModel.get_best_fit_parameters — Function

get_best_fit_parameters(results; chains=nothing, burn_in::Float64=0.2, with_errors::Bool=false, thin::Int64=1, print_table::Bool=false)

Get the best-fit parameters from the optimization or MCMC results.

Arguments

results::Union{Vector{@NamedTuple{θ::Vector{Float64}, f::Float64}}, Matrix{Float64}}: The optimization or MCMC results.

If results::Vector{@NamedTuple{θ::Vector{Float64}, f::Float64}}: It is assumed to be from the optimizer results, where each element is a named tuple containing the parameter vector (θ) and the χ² values (f).

If results::Matrix{Float64}: It is assumed to be the χ² matrix corresponding to MCMC chains.

Keyword Arguments

chains::Union{Array{Float64, 3}, Nothing}=nothing: MCMC chains. The dimensions should be (n_steps, n_chains, n_parameters). This is only required when results is a Matrix{Float64}.

burn_in::Float64=0.2: The fraction of MCMC chains to discard as burn-in (default: 0.2).

with_errors::Bool=false: If true, then also return 1-sigma uncertainties for best-fit parameters. But it will only work if chains are provided.

thin::Int64=1: The thinning factor to apply to the MCMC chains while calculating parameter uncertainties.

print_table::Bool=false: If true, then print the best-fit parameters along with [-1σ, +1σ] and [-2σ, +2σ] ranges in a table format. This option is only available when with_errors=true.

Returns

best_θ::Vector{Float64}: Best-fit parameter values.

log_posterior::Float64: Log-posterior value corresponding to the best-fit parameters.

lower_err::Vector{Float64}: Lower 1-sigma uncertainties (only if with_errors=true).

upper_err::Vector{Float64}: Upper 1-sigma uncertainties (only if with_errors=true).

LensFactory.LensModel.get_best_model — Function

get_best_model(model::ModelConfig, chains::Array{Float64, 3}, logL::Matrix{Float64})

Get the best-fit lens model based on the MCMC results stored in chains and logL. The best-fit parameters are determined by the minimum log-likelihood in logL.

Arguments

model::ModelConfig: The lens model configuration used for the MCMC fit.

chains::Array{Float64, 3}: The MCMC chains containing the sampled parameter values. The dimensions should be (nsteps, nchains, n_parameters).

logL::Matrix{Float64}: The log-likelihood values corresponding to each sample in the MCMC chains. The dimensions should be (nsteps, nsteps).

Returns

best_model: Best-fit lens model constructed using the best-fit parameters.

logL: Log-likelihoohood correspodning to best-fit lens model.

LensFactory.LensModel.get_potential — Function

get_potential(best_model::Lenses.AbstractLens, θx::T, θy::T;
              reference::Tuple{Float64, Float64}) where T <: Union{RV, ROA}

get_potential(jld2_file::JLD2.JLDFile, θx::T, θy::T; unit::Symbol=:RA_DEC) where T <: Union{ROA, Vector{Int64}}

Calculate the lensing potential at the given coordinates (θx, θy) for the best-fit model. The user can specify the unit of the input coordinates as either RA/DEC or arcseconds.

Arguments

jld2_file::JLD2.JLDFile or best_model::Lenses.AbstractLens: JLD2 data file containing the fit results or the best-fit lens model.

θx: x-coordinates

θy: y-coordinates

Returns

ψ: Lensing potential at the input coordinates for the best-fit model.

LensFactory.LensModel.get_deflection — Function

get_deflection(best_model::Lenses.AbstractLens, θx::T, θy::T; unit::Symbol=:RA_DEC) where T <: Union{RV, ROA}

get_deflection(jld2_file::JLD2.JLDFile, θx::T, θy::T; unit::Symbol=:RA_DEC) where T <: Union{RV, ROA}

Calculate the lensing deflection at the given coordinates (θx, θy) for the best-fit model. The user can specify the unit of the input coordinates as either RA/DEC or arcseconds.

Arguments

jld2_file::JLD2.JLDFile or best_model::Lenses.AbstractLens: JLD2 data file containing the fit results or the best-fit lens model.

θx: x-coordinates

θy: y-coordinates

Keyword Arguments

reference = (0.0, 0.0): Reference (RA, Dec) coordinate (in degrees).

Returns

αx: x-component of the deflection angle (in arcseconds).

αy: y-component of the deflection angle (in arcseconds).

LensFactory.LensModel.get_jacobian — Function

get_jacobian(best_model::Lenses.AbstractLens, θx::T, θy::T; unit::Symbol=:RA_DEC) where T <: Union{RV, ROA}

get_jacobian(jld2_file::JLD2.JLDFile, θx::T, θy::T; unit::Symbol=:RA_DEC) where T <: Union{RV, ROA}

Calculate the Jacobian (i.e., deformation tensor) at the given coordinates (θx, θy) for the best-fit model. The user can specify the unit of the input coordinates as either RA/DEC or arcseconds.

Arguments

jld2_file::JLD2.JLDFile or best_model::Lenses.AbstractLens: JLD2 data file containing the fit results or the best-fit lens model.

θx: x-coordinates

θy: y-coordinates

Keyword Arguments

reference = (0.0, 0.0): Reference (RA, Dec) coordinate (in degrees).

Returns

ψxx: xx-component of the jacobian.

ψyy: yy-component of the jacobian.

ψxy: xy-component of the jacobian.

LensFactory.LensModel.predict_image — Function

predict_image(jld2_file::JLD2.JLDFile, θx::T, θy::T, z_s::Float64; unit::Symbol=:RA_DEC) where T <: Union{RV, Vector{Float64}}

Predict counter-image positions, magnifications and time delays based on the best-fit lens model. The function can take either a single observed image or multiple observed images of the same system. If multiple images are provided then the function will calculate the barycentric source position and then predict the counter-image positions, magnifications and time delays.

Arguments

data_jld2::JLD2.JLDFile: JLD2 data file containing the fit results.

θx: x-coordinate(s) of the observed image position(s).

θy: y-coordinate(s) of the observed image position(s).

z_s::Float64: Source redshift.

Keyword Arguments

unit::Symbol=:RA_DEC: Unit of the input coordinates.

:RA_DEC: (θx, θy) are assumed to be in RA/DEC (in degrees).

:arcsec: (θx, θy) are assumed to be in arcseconds.

Returns

nothing: Prints the counter-image positions, magnifications and time delays in a table format. The table is sorted based on the time delay and at the end also contains the best-fit model predicted source position.

LensFactory.LensModel.save_best_fits — Function

save_best_fits(file_name::String; save_potential::Bool=true, save_deflection::Bool=true, save_kappa::Bool=true, save_gamma::Bool=true)

Save fits files for the best-fit model based on the MCMC results stored in file_name. The user can choose which lensing quantities to save by setting the corresponding boolean flags.

Arguments

file_name::String: Path to the JLD2 file containing the MCMC results

Keyword Arguments

save_potential::Bool=true: Whether to save the lensing potential as potential.fits

save_deflection::Bool=true: Whether to save the deflection angles as alpha_x.fits and alpha_y.fits

save_kappa::Bool=true: Whether to save the convergence map as kappa.fits

save_gamma::Bool=true: Whether to save the shear maps as gamma1.fits and gamma2.fits

Returns

nothing: Saves FITS files to disk

LensFactory.LensModel.get_best_fit_rms — Function

get_best_fit_rms(model::ModelConfig, chains::Array{Float64, 3}, logL::Matrix{Float64}; burn_in::Float64=0.2)

Calculate the RMS of the best-fit model based on the MCMC results stored in chains and logL. The function prints a table showing the RMS for each knot image, as well as a global total RMS. If

Arguments

model::ModelConfig: The lens model configuration used for the MCMC fit.

chains::Array{Float64, 3}: The MCMC chains containing the sampled parameter values. The dimensions should be (nsteps, nchains, n_parameters).

logL::Matrix{Float64}: The logL values corresponding to each sample in the MCMC chains. The dimensions should be (nsteps, nsteps).

Keyword Arguments

burn_in::Float64=0.2: The fraction of MCMC chains to discard as burn-in while determining best-fit lens model.

Returns

nothing: Prints a table to the console with the RMS for each knot image, as well as total RMS.

LensFactory.LensModel.get_AIC — Function

get_AIC(model::ModelConfig, chains::Array{Float64,3}, logL::Matrix{Float64};
        burn_in::Float64 = 0.2)

Compute the Akaike Information Criterion (AIC) for a fitted LensFactory model.

\[ \rm{AIC} = -2 \ln L + 2 k\]

Arguments

model : ModelConfig from read_input

logL : Log-likelihood matrix of shape (nsteps, nchains) from fit_model

Keyword Arguments

burn_in : Fraction of chain to discard as burn-in

Returns

Float64: The AIC value for the model.

LensFactory.LensModel.get_BIC — Function

get_BIC(model::ModelConfig, chains::Array{Float64,3}, logL::Matrix{Float64}, n_data::Int; 
        burn_in::Float64 = 0.2)

Compute the Bayesian Information Criterion (BIC) for a fitted LensFactory model.

\[ \rm{BIC} = -2 \ln L + k \ln(n)\]

Arguments

model : ModelConfig from read_input

logL : Log-likelihood matrix of shape (nsteps, nchains) from fit_model

n_data : Total number of observed constraints

burn_in : Fraction of chain to discard as burn-in (default 0.2)