LensModel

The LensModel module is the interface between the user and

LensFactory.LensModel.LensModelIO.read_input — Function

read_input(filename::AbstractString) -> ModelConfig

Reads the input YAML file and constructs a ModelConfig struct containing all the necessary information for lens modeling and sampling. For details on the expected structure of the input YAML file, please refer to Example - 2.

Arguments

filename::AbstractString: Path to the input YAML file.

Returns

ModelConfig::ModelConfig: A struct containing the observation details, cosmology, lens

configuration, source configuration, parameter definitions, and sampling configuration.

LensFactory.LensModel.LensModelFit.fit_model — Function

fit_model(model::ModelConfig; save::Bool=true, file_name::Union{String, Nothing}=nothing)

Performs lens model fitting using the given model.

Arguments:

model::ModelConfig: The lens model configuration containing the observation, lens, source, and sampler details.

save::Bool=true: Whether to save the MCMC results to a JLD2 file (default: true)

file_name::Union{String, Nothing}=nothing: The name of the JLD2 file to save results to. If no name is provided then a name will be generated based on the lens name and current date (i.e., LensName_DDMMYYYY.jld2).

Returns:

chains::Array{Float64, 3}: The MCMC chains containing sampled parameter values

chi2::Matrix{Float64}: The chi-squared values corresponding to each sample in the chains

LensFactory.LensModel.save_best_fit — Function

save_best_fit(file_name::String; save_potential::Bool=true, save_deflection::Bool=true, save_kappa::Bool=true, save_gamma::Bool=true)

Save fits files for the best-fit model based on the MCMC results stored in file_name. The user can choose which lensing quantities to save by setting the corresponding boolean flags.

Arguments

file_name::String: Path to the JLD2 file containing the MCMC results

Keyword Arguments

save_potential::Bool=true: Whether to save the lensing potential as potential.fits

save_deflection::Bool=true: Whether to save the deflection angles as alpha_x.fits and alpha_y.fits

save_kappa::Bool=true: Whether to save the convergence map as kappa.fits

save_gamma::Bool=true: Whether to save the shear maps as gamma1.fits and gamma2.fits

Returns

Nothing (saves FITS files to disk)

LensFactory.LensModel.get_potential — Function

get_potential(file_name::String, θx::T, θy::T; unit::Symbol=:RA_DEC) where T <: Union{ROA, Vector{Int64}}

Calculate the lensing potential at the given coordinates (θx, θy) for the best-fit model based on the MCMC results stored in file_name. The user can specify the unit of the input coordinates as either RA/DEC or arcseconds.

Arguments

file_name::String: Path to the JLD2 file containing the MCMC results

θx: x-coordinates

θy: y-coordinates

Keyword Arguments

unit::Symbol=:RA_DEC: Unit of the input coordinates. Supported units are :RA_DEC and :arcsec. If :RA_DEC then (θx, θy) = (RA, DEC) is assumed and will be converted to arcseconds relative to the reference position in the model.

Returns

ψ: Lensing potential at the input coordinates for the best-fit model.

LensFactory.LensModel.get_deflection — Function

get_deflection(file_name::String, θx::T, θy::T; unit::Symbol=:RA_DEC) where T <: Union{ROA, Vector{Int64}}

Calculate the lensing deflection at the given coordinates (θx, θy) for the best-fit model based on the MCMC results stored in file_name. The user can specify the unit of the input coordinates as either RA/DEC or arcseconds.

Arguments

file_name::String: Path to the JLD2 file containing the MCMC results

θx: x-coordinates

θy: y-coordinates

Keyword Arguments

unit::Symbol=:RA_DEC: Unit of the input coordinates. Supported units are :RA_DEC and :arcsec. If :RA_DEC then (θx, θy) = (RA, DEC) is assumed and will be converted to arcseconds relative to the reference position in the model.

Returns

αx: x-component of the deflection angle (in arcseconds).

αy: y-component of the deflection angle (in arcseconds).

LensFactory.LensModel.get_jacobian — Function

get_jacobian(file_name::String, θx::T, θy::T; unit::Symbol=:RA_DEC) where T <: Union{ROA, Vector{Int64}}

Calculate the Jacobian (i.e., deformation tensor) of the lens mapping at the given coordinates (θx, θy) for the best-fit model based on the MCMC results stored in file_name. The user can specify the unit of the input coordinates as either RA/DEC or arcseconds.

Arguments

file_name::String: Path to the JLD2 file containing the MCMC results

θx: x-coordinates

θy: y-coordinates

Keyword Arguments

unit::Symbol=:RA_DEC: Unit of the input coordinates. Supported units are :RA_DEC and :arcsec. If :RA_DEC then (θx, θy) = (RA, DEC) is assumed and will be converted to arcseconds relative to the reference position in the model.

Returns

ψxx: xx-component of the jacobian.

ψyy: yy-component of the jacobian.

ψxy: xy-component of the jacobian.