LensModel

The LensModel module is the interface between the user and

LensFactory.LensModel.read_input — Function

read_input(input_filename::AbstractString)

Reads the input YAML file and constructs a ModelConfig struct containing all the necessary information for lens modeling and sampling. For details on the expected structure of the input YAML file, please refer to Example - 2.

Arguments

input_filename::AbstractString: Path to the input YAML file.

Returns

ModelConfig::ModelConfig: A struct containing the observation details, cosmology, lens configuration, source configuration, parameter definitions, and sampling configuration.

LensFactory.LensModel.fit_model — Function

fit_model(model::ModelConfig; save::Bool=true, file_name::Union{String, Nothing}=nothing)

Performs lens model fitting using the given model.

Arguments

model::ModelConfig: The lens model configuration containing the observation, lens, source, and sampler details.

save::Bool=true: Whether to save the MCMC results to a JLD2 file (default: true)

file_name::Union{String, Nothing}=nothing: The name of the JLD2 file to save results. If no name is provided then a name will be generated based on the lens name and current date (i.e., LensName_DDMMYYYY.jld2).

Returns

chains::Array{Float64, 3}: The MCMC chains containing sampled parameter values

log_posterior::Matrix{Float64}: The log-posterior values corresponding to each sample in the chains

LensFactory.LensModel.free_parameter_names — Function

free_parameter_names(model::ModelConfig)

Get a vector of tuples containing the owner and name of each free parameter in the model.

Arguments

model::ModelConfig: The lens model configuration containing the observation, lens, source, and sampler details.

Returns

Vector{Tuple{Symbol, Symbol}}: A vector of tuples containing the owner and name of each free parameter.

LensFactory.LensModel.get_best_fit_parameters — Function

get_best_fit_parameters(results; chains=nothing, burn_in::Float64=0.2, with_errors::Bool=false, thin::Int64=1, print_table::Bool=false)

Get the best-fit parameters from the optimization or MCMC results.

Arguments

results::Union{Vector{@NamedTuple{θ::Vector{Float64}, f::Float64}}, Matrix{Float64}}: The optimization or MCMC results.

If results::Vector{@NamedTuple{θ::Vector{Float64}, f::Float64}}: It is assumed to be from the optimizer results, where each element is a named tuple containing the parameter vector (θ) and the χ² values (f).

If results::Matrix{Float64}: It is assumed to be the χ² matrix corresponding to MCMC chains.

Keyword Arguments

chains::Union{Array{Float64, 3}, Nothing}=nothing: MCMC chains. The dimensions should be (n_steps, n_chains, n_parameters). This is only required when results is a Matrix{Float64}.

burn_in::Float64=0.2: The fraction of MCMC chains to discard as burn-in (default: 0.2).

with_errors::Bool=false: If true, then also return 1-sigma uncertainties for best-fit parameters. But it will only work if chains are provided.

thin::Int64=1: The thinning factor to apply to the MCMC chains while calculating parameter uncertainties.

print_table::Bool=false: If true, then print the best-fit parameters along with [-1σ, +1σ] and [-2σ, +2σ] ranges in a table format. This option is only available when with_errors=true.

Returns

best_θ::Vector{Float64}: Best-fit parameter values.

log_posterior::Float64: Log-posterior value corresponding to the best-fit parameters.

lower_err::Vector{Float64}: Lower 1-sigma uncertainties (only if with_errors=true).

upper_err::Vector{Float64}: Upper 1-sigma uncertainties (only if with_errors=true).

LensFactory.LensModel.get_best_model — Function

get_best_model(model::ModelConfig, chains::Array{Float64, 3}, logL::Matrix{Float64})

Get the best-fit lens model based on the MCMC results stored in chains and logL. The best-fit parameters are determined by the minimum log-likelihood in logL.

Arguments

model::ModelConfig: The lens model configuration used for the MCMC fit.

chains::Array{Float64, 3}: The MCMC chains containing the sampled parameter values. The dimensions should be (nsteps, nchains, n_parameters).

logL::Matrix{Float64}: The log-likelihood values corresponding to each sample in the MCMC chains. The dimensions should be (nsteps, nsteps).

Returns

best_model: Best-fit lens model constructed using the best-fit parameters.

logL: Log-likelihoohood correspodning to best-fit lens model.

LensFactory.LensModel.get_potential — Function

get_potential(best_model::Lenses.AbstractLens, θx::T, θy::T; unit::Symbol=:RA_DEC) where T <: Union{RV, ROA}

get_potential(jld2_file::JLD2.JLDFile, θx::T, θy::T; unit::Symbol=:RA_DEC) where T <: Union{ROA, Vector{Int64}}

Calculate the lensing potential at the given coordinates (θx, θy) for the best-fit model. The user can specify the unit of the input coordinates as either RA/DEC or arcseconds.

Arguments

jld2_file::JLD2.JLDFile or best_model::Lenses.AbstractLens: JLD2 data file containing the fit results or the best-fit lens model.

θx: x-coordinates

θy: y-coordinates

Keyword Arguments

unit::Symbol=:RA_DEC: Unit of the input coordinates.

:RA_DEC: (θx, θy) are assumed to be in RA/DEC (in degrees).

:arcsec: (θx, θy) are assumed to be in arcseconds.

Returns

ψ: Lensing potential at the input coordinates for the best-fit model.

LensFactory.LensModel.get_deflection — Function

get_deflection(best_model::Lenses.AbstractLens, θx::T, θy::T; unit::Symbol=:RA_DEC) where T <: Union{RV, ROA}

get_deflection(jld2_file::JLD2.JLDFile, θx::T, θy::T; unit::Symbol=:RA_DEC) where T <: Union{RV, ROA}

Calculate the lensing deflection at the given coordinates (θx, θy) for the best-fit model. The user can specify the unit of the input coordinates as either RA/DEC or arcseconds.

Arguments

jld2_file::JLD2.JLDFile or best_model::Lenses.AbstractLens: JLD2 data file containing the fit results or the best-fit lens model.

θx: x-coordinates

θy: y-coordinates

Keyword Arguments

unit::Symbol=:RA_DEC: Unit of the input coordinates.

:RA_DEC: (θx, θy) are assumed to be in RA/DEC (in degrees).

:arcsec: (θx, θy) are assumed to be in arcseconds.

Returns

αx: x-component of the deflection angle (in arcseconds).

αy: y-component of the deflection angle (in arcseconds).

LensFactory.LensModel.get_jacobian — Function

get_jacobian(best_model::Lenses.AbstractLens, θx::T, θy::T; unit::Symbol=:RA_DEC) where T <: Union{RV, ROA}

get_jacobian(jld2_file::JLD2.JLDFile, θx::T, θy::T; unit::Symbol=:RA_DEC) where T <: Union{RV, ROA}

Calculate the Jacobian (i.e., deformation tensor) at the given coordinates (θx, θy) for the best-fit model. The user can specify the unit of the input coordinates as either RA/DEC or arcseconds.

Arguments

jld2_file::JLD2.JLDFile or best_model::Lenses.AbstractLens: JLD2 data file containing the fit results or the best-fit lens model.

θx: x-coordinates

θy: y-coordinates

Keyword Arguments

unit::Symbol=:RA_DEC: Unit of the input coordinates.

:RA_DEC: (θx, θy) are assumed to be in RA/DEC (in degrees).

:arcsec: (θx, θy) are assumed to be in arcseconds.

Returns

ψxx: xx-component of the jacobian.

ψyy: yy-component of the jacobian.

ψxy: xy-component of the jacobian.

LensFactory.LensModel.predict_image — Function

predict_image(jld2_file::JLD2.JLDFile, θx::T, θy::T, z_s::Float64; unit::Symbol=:RA_DEC) where T <: Union{RV, Vector{Float64}}

Predict counter-image positions, magnifications and time delays based on the best-fit lens model. The function can take either a single observed image or multiple observed images of the same system. If multiple images are provided then the function will calculate the barycentric source position and then predict the counter-image positions, magnifications and time delays.

Arguments

data_jld2::JLD2.JLDFile: JLD2 data file containing the fit results.

θx: x-coordinate(s) of the observed image position(s).

θy: y-coordinate(s) of the observed image position(s).

z_s::Float64: Source redshift.

Keyword Arguments

unit::Symbol=:RA_DEC: Unit of the input coordinates.

:RA_DEC: (θx, θy) are assumed to be in RA/DEC (in degrees).

:arcsec: (θx, θy) are assumed to be in arcseconds.

Returns

nothing: Prints the counter-image positions, magnifications and time delays in a table format. The table is sorted based on the time delay and at the end also contains the best-fit model predicted source position.

LensFactory.LensModel.save_best_fits — Function

save_best_fits(file_name::String; save_potential::Bool=true, save_deflection::Bool=true, save_kappa::Bool=true, save_gamma::Bool=true)

Save fits files for the best-fit model based on the MCMC results stored in file_name. The user can choose which lensing quantities to save by setting the corresponding boolean flags.

Arguments

file_name::String: Path to the JLD2 file containing the MCMC results

Keyword Arguments

save_potential::Bool=true: Whether to save the lensing potential as potential.fits

save_deflection::Bool=true: Whether to save the deflection angles as alpha_x.fits and alpha_y.fits

save_kappa::Bool=true: Whether to save the convergence map as kappa.fits

save_gamma::Bool=true: Whether to save the shear maps as gamma1.fits and gamma2.fits

Returns

nothing: Saves FITS files to disk

LensFactory.LensModel.check_parity — Function

check_parity(model::ModelConfig, chains::Array{Float64, 3}, logL::Matrix{Float64})

Check the parity of the best-fit model against the input parities for each knot image. This function assumes that parity was enforced during the modelling (i.e., model.source_config.use_parity = true). If parity was not enforced, an error is thrown. The function prints a table comparing the input parity and the best-fit model parity for each knot image, along with a status indicating whether the parity matches or not.

Arguments

data_jld2::JLD2.JLDFile: JLD2 data file containing the fit results.

Returns

nothing: Prints a table to the console with input and best-fit model parities for each knot image.

LensFactory.LensModel.get_best_fit_rms — Function

get_best_fit_rms(model::ModelConfig, chains::Array{Float64, 3}, logL::Matrix{Float64}; check_parity::Bool=false, burn_in::Float64=0.2)

Calculate the RMS of the best-fit model based on the MCMC results stored in chains and logL. The function prints a table showing the RMS for each knot image, as well as a global total RMS. If check_parity is set to true, the function will also check the parity of each knot image and print a warning if any parity does not match.

Arguments

model::ModelConfig: The lens model configuration used for the MCMC fit.

chains::Array{Float64, 3}: The MCMC chains containing the sampled parameter values. The dimensions should be (nsteps, nchains, n_parameters).

logL::Matrix{Float64}: The logL values corresponding to each sample in the MCMC chains. The dimensions should be (nsteps, nsteps).

Keyword Arguments

check_parity::Bool=false: Whether to check the parity of each knot image against the input parity.

burn_in::Float64=0.2: The fraction of MCMC chains to discard as burn-in while determining best-fit lens model.

Returns

nothing: Prints a table to the console with the RMS for each knot image, as well as total RMS.