Lenses

LensFactory.Lenses.get_meshgrid — Function

get_meshgrid(θx::RV, θy::RV, dθ::RV) --> Tuple{Matrix{Float64}, Matrix{Float64}}

Generate a meshgrid of coordinates on which various quantities can be evaluated. At present, this function only generates square pixels. In future if the need arises, it can be extended to generate rectangular pixels as well.

LensFactory.Lenses.get_critical_density — Function

get_critical_density(D_d::Float64, adis::Float64; unit::Symbol=:kg_m2) --> Float64

get_critical_density(D_d::Float64, D_ds::Float64, D_s::Float64s; unit::Symbol=:kg_m2) --> Float64

Calculate the critical surface density,

\[Σ_{\rm cr} = \frac{c^2}{4 π {\rm G}} \frac{D_s}{D_d D_{ds}},\]

given the angular diameter distances. The result can be returned in different units,

:kg_m2 $\Rightarrow{\rm kg/m^2}$,

:msun_pc2 $\Rightarrow{\rm M_⊙/pc^2}$,

:msun_arcsec2 $\Rightarrow{\rm M_⊙/arcsec^2}$.

LensFactory.Lenses.get_potential — Function

get_potential(lens::AbstractLens, θx::T, θy::T) where T <: RV --> RV

get_potential(lens::AbstractLens, θx::T, θy::T) where T <: ROA --> ROA

LensFactory.Lenses.get_deflection — Function

get_deflection(lens::AbstractLens, θx::T, θy::T) where T <: RV --> Tuple{RV, RV}

get_deflection(lens::AbstractLens, θx::T, θy::T) where T <: ROA --> Tuple{ROA, ROA}

Calculates the deflection angles (i.e., the gradient of the potential) for a given lens model. Returns a tuple of deflection components, i.e., $(ψ_x, ψ_y)$.

LensFactory.Lenses.get_jacobian — Function

get_jacobian(lens::AbstractLens, θx::T, θy::T) where T <: RV --> Tuple{RV, RV, RV}

get_jacobian(lens::AbstractLens, θx::T, θy::T) where T <: ROA --> Tuple{ROA, ROA, ROA}

Calculates the jacobian (i.e., deformation tensor) of the lens mapping for a given lens model. Since the jacobian is symmetric (for single lens plane), only three components are returned, i.e., $(ψ_{xx}, ψ_{yy}, ψ_{xy})$.

LensFactory.Lenses.get_time_delay — Function

get_time_delay(lens::AbstractLens, θx::T, θy::T, adis::Float64, z_d::RV, D_d::RV, β::NTuple{2, RV}) where T <: RV --> RV

get_time_delay(lens::AbstractLens, θx::T, θy::T, adis::Float64, z_d::RV, D_d::RV, β::NTuple{2, RV}) where T <: ROA --> ROA

Calculates the time delay for a given lens model. The corresponding expression is given as,

\[t_d(\pmb{θ}; \pmb{β}) = \frac{1+z_l}{\rm c} \frac{D_d D_s}{D_{ds}} \theta_0^2 \left[ \frac{(\pmb{θ} - \pmb{β})^2}{2} - \frac{D_{ds}}{D_s} \psi(\pmb{θ}) \right],\]

where $\theta_0$ is normalizing angular unit.

LensFactory.Lenses.get_magnification_image — Function

get_magnification_image(lens::AbstractLens, θx::T, θy::T) where T <: RV

get_magnification_image(lens::AbstractLens, θx::T, θy::T) where T <: ROA

LensFactory.Lenses.get_image — Function

get_image(lens::AbstractLens, θx::ROA, θy::ROA, adis::Float64, β::NTuple{2, RV}) --> Vector{NTuple{2, RV}}

LensFactory.Lenses.get_critical_curve — Function

get_critical_curve(lens::AbstractLens, θx::T, θy::T, adis::Float64) where T <: Matrix{<:RV}

LensFactory.Lenses.get_caustic — Function

get_caustic(lens::AbstractLens, θx::T, θy::T, adis::Float64) where T <: Matrix{<:RV}

LensFactory.Lenses.get_critical_area — Function

get_critical_area(lens::AbstractLens, θx::T, θy::T, adis::Float64) where T <: Matrix{<:RV} --> Float64

LensFactory.Lenses.get_einstein_angle — Function

get_einstein_angle(lens::AbstractLens, θx::T, θy::T, adis::Float64) where T <: Matrix{<:RV} --> Float64