LensFactory.LensModel.plot_corner — Function

LensFactory.LensModel.plot_corner(chains, chi2)

Generates a corner plot for the given MCMC chains and chi-squared values.

Arguments

chains: MCMC chains of shape (nsteps, nchains, n_params).

chi2: Chi-squared values corresponding to the chains, of shape (nsteps, nchains).

Keyword arguments

param_names=nothing: Optional list of parameter names for labeling the axes.

burn_in::Float64 = 0.3: Fraction of the initial samples to discard as burn-in.

thinning::Int64 = 100: Interval for thinning the chains to reduce autocorrelation.

save_plot::Bool = true: Whether to save the plot as "corner.png".

plot_name::String = "./corner.png": Filename for saving the plot.

resolution::Int64 = 2: Resolution for saving the plot.

Returns

A Makie figure object containing the corner plot.

LensFactory.LensModel.plot_corner(results)

Generates a corner plot for the optimizer results. This plot is useful to check the convergence of the optimizer runs. For example, if the optimizer is run for 1000 times, the corner plot will show the distribution of the parameters in each of the converged runs.

Arguments

results: Vector of optimizer results, each containing .θ (parameter vector) and .f (chi-squared value).

Keyword arguments

param_names = nothing: List of parameter names for labeling the axes.

save_plot::Bool = true: Whether to save the plot as "corner_optimizer.png".

plot_name::String = "./corner_optimizer.png": Filename for saving the plot.

resolution::Int64 = 2: Resolution for saving the plot.

Returns

A Makie figure object containing the corner plot.

LensFactory.LensModel.plot_trace — Function

LensFactory.LensModel.plot_trace(chains)

Generates trace plots for the given MCMC chains.

Arguments

chains: MCMC chains of shape (nsteps, nchains, n_params).

Keyword arguments

param_names = nothing: Optional list of parameter names for labeling the axes.

burn_in::Float64 = 0.0: Fraction of the initial samples to discard as burn-in.

thinning::Int64 = 1: Interval for thinning the chains to reduce autocorrelation.

save_plot::Bool = true: Whether to save the plot as "corner_optimizer.png".

plot_name::String = "./trace.png": Filename for saving the plot.

resolution::Int64 = 2: Resolution for saving the plot.

Returns

A Makie figure object containing the trace plots.

LensFactory.LensModel.plot_best_model — Function

LensFactory.LensModel.plot_best_model(model, chains, chi2)

Generates a plot comparing the observed image positions with the predicted image positions from the best-fit lens model.

Arguments

model: The lens model configuration.

chains: MCMC chains of shape (nsteps, nchains, n_params).

chi2: Chi-squared values corresponding to the chains, of shape (nsteps, nchains).

Keyword arguments

z_s::RV = 1.5: Source redshift.

plot_critical::Bool = true: Whether to plot the critical curves.

critical_tan_kws::NamedTuple = (color=:red, linewidth=2, linestyle=:solid): Keyword arguments for plotting the tangential critical curve.

critical_rad_kws::NamedTuple = (color=:red, linewidth=2, linestyle=:dash): Keyword arguments for plotting the radial critical curve.

plot_caustics::Bool = true: Whether to plot the caustics.

caustic_tan_kws::NamedTuple = (color=:green, linewidth=2, linestyle=:solid): Keyword arguments for plotting the tangential caustic.

caustic_rad_kws::NamedTuple = (color=:green, linewidth=2, linestyle=:dash): Keyword arguments for plotting the radial caustic.

save_plot::Bool = true: Whether to save the plot as "best_model.png".

plot_name::String = "./best_model.png": Filename for saving the plot.

resolution::Int64 = 2: Resolution for saving the plot.

Returns

A Makie figure object containing the comparison plot.