LensFactory.Lenses.init_SersicLens — Type

init_SersicLens(D_d::RV=NaN, x_c::RV=0.0, y_c::RV=0.0, mass::RV=NaN, x_e::RV=NaN, n::RV=4)

Initialize a Sersic lens with the given parameters.

Keyword Arguments

D_d::RV = NaN: ADD from observer to lens (in $\rm \mathbf{meters}$).

x_c::RV = 0.0: x-coordinate of the lens (in $\rm \mathbf{arcseconds}$).

y_c::RV = 0.0: y-coordinate of the lens (in $\rm \mathbf{arcseconds}$).

mass::RV= NaN: Mass of the lens (in $\rm \mathbf{M_\odot}$).

x_e::RV = NaN: Half-mass radius (in $\rm \mathbf{arcseconds}$).

n::RV = 4.0: Sersic index.

LensFactory.Lenses.SersicLens.potential! — Function

potential!(ψ::T, θx::T, θy::T, D_d::RV, θxc::RV, θyc::RV, mass::RV, θe::RV, n::RV) where T <: RV

potential!(ψ::T, θx::T, θy::T, D_d::RV, θxc::RV, θyc::RV, mass::RV, θe::RV, n::RV) where T <: ROA

Calculate potential at given coordinates for a Sersic lens and update the potential in place.

Arguments

ψ: Potential at given coordinates

θx: x-coordinate(s) (in $\rm \mathbf{arcseconds}$).

θy: y-coordinate(s) (in $\rm \mathbf{arcseconds}$).

D_d::RV: ADD from the observer to the lens (in $\rm \mathbf{meters}$).

θxc::RV: x-coordinate of the lens (in $\rm \mathbf{arcseconds}$).

θyc::RV: y-coordinate of the lens (in $\rm \mathbf{arcseconds}$).

mass::RV: Mass of the lens (in $\rm \mathbf{M_\odot}$).

θe::RV: Half-light radius of the lens (in $\rm \mathbf{arcseconds}$).

n::RV: Sersic index of the lens.

LensFactory.Lenses.SersicLens.deflection! — Function

deflection!(ψx::T, ψy::T, θx::T, θy::T, D_d::RV, θxc::RV, θyc::RV, mass::RV, θe::RV, n::RV) where T <: RV

deflection!(ψx::T, ψy::T, θx::T, θy::T, D_d::RV, θxc::RV, θyc::RV, mass::RV, θe::RV, n::RV) where T <: ROA

Calculate deflection at given coordinates for a Sersic lens and update the deflection in place.

Arguments

ψx: x-component of deflection at given coordinates

ψy: y-component of deflection at given coordinates

θx: x-coordinate(s) (in $\rm \mathbf{arcseconds}$).

θy: y-coordinate(s) (in $\rm \mathbf{arcseconds}$).

D_d::RV: ADD from the observer to the lens (in $\rm \mathbf{meters}$).

θxc::RV: x-coordinate of the lens (in $\rm \mathbf{arcseconds}$).

θyc::RV: y-coordinate of the lens (in $\rm \mathbf{arcseconds}$).

mass::RV: Mass of the lens (in $\rm \mathbf{M_\odot}$).

θe::RV: Half-light radius of the lens (in $\rm \mathbf{arcseconds}$).

n::RV: Sersic index of the lens.

LensFactory.Lenses.SersicLens.jacobian! — Function

jacobian!(ψxx::T, ψyy::T, ψxy::T, θx::T, θy::T, D_d::RV, θxc::RV, θyc::RV, mass::RV, θe::RV, n::RV) where T <: RV

jacobian!(ψxx::T, ψyy::T, ψxy::T, θx::T, θy::T, D_d::RV, θxc::RV, θyc::RV, mass::RV, θe::RV, n::RV) where T <: ROA

Calculate Jacobian of the deflection at given coordinates for a Sersic lens and update the Jacobian in place.

Arguments

ψxx: x-component of Jacobian at given coordinates

ψyy: y-component of Jacobian at given coordinates

ψxy: xy-component of Jacobian at given coordinates

θx: x-coordinate(s) (in $\rm \mathbf{arcseconds}$).

θy: y-coordinate(s) (in $\rm \mathbf{arcseconds}$).

D_d::RV: ADD from the observer to the lens (in $\rm \mathbf{meters}$).

θxc::RV: x-coordinate of the lens (in $\rm \mathbf{arcseconds}$).

θyc::RV: y-coordinate of the lens (in $\rm \mathbf{arcseconds}$).

mass::RV: Mass of the lens (in $\rm \mathbf{M_\odot}$).

θe::RV: Half-light radius of the lens (in $\rm \mathbf{arcseconds}$).

n::RV: Sersic index of the lens.

LensFactory.Lenses.SersicLens.scale_to_halflight — Function

scale_to_halflight(;n::RV=NaN, x_s::RV=NaN)

Calculate the scale radius of a Sersic lens given the half-light radius and the Sersic index.

Keyword Arguments

n::RV = NaN: Sersic index of the lens.

x_s::RV = NaN: Half-light radius of the lens (in $\rm \mathbf{arcseconds}$).

Returns

θs: Scale radius of the lens (in $\rm \mathbf{arcseconds}$).